Split libreimu to libreimu and libreimu_fdt
