+++ meta-mcst/meta-common/recipes-mcst/hwmon-instantiator/files/Makefile
@@ -3,7 +3,7 @@ BINARYNAME=hwmon-instantiator
 all: $(BINARYNAME)
 
 $(BINARYNAME): $(BINARYNAME).o
-	$(CC) -o $(BINARYNAME) $(BINARYNAME).o -lfdt -lreimu $(LDFLAGS)
+	$(CC) -o $(BINARYNAME) $(BINARYNAME).o -lfdt -lreimu -lreimu_fdt $(LDFLAGS)
 
 $(BINARYNAME).o: $(BINARYNAME).c
 	$(CC) -o $(BINARYNAME).o -c $(BINARYNAME).c
